pull up summary formatting to util.js file. move summary.json writing to specific "reporter" functions
