+++ lib/console.js
@@ -1,23 +1,11 @@
 var table = require('cli-table');
-
-module.exports = function (context, instances) {
-  let totals = {total: 0, pass: 0, fail: 0};
-  instances.forEach(instance => {
-    let instanceSummary = {total: 0, pass: 0, fail: 0};
-    instance.testResults.forEach(test => {
-      instanceSummary.total = instanceSummary.total + 1;
-      instanceSummary.pass = (test.state === 'passed') ? instanceSummary.pass + 1 : instanceSummary.pass;
-      instanceSummary.fail = (test.state === 'failed') ? instanceSummary.fail + 1 : instanceSummary.fail;
-    });
-    instance.summary = instanceSummary;
-    totals.total = totals.total + instanceSummary.total;
-    totals.pass = totals.pass + instanceSummary.pass;
-    totals.fail = totals.fail + instanceSummary.fail;
-  });
+var util = require('./util');
+module.exports = function (context, _instances) {
+  let output = util.formatSummary(_instances);
   let tabl = new table({
     head: ['tags', 'pass', 'fail', 'total', 'report']
   });
-  instances.forEach(function (instance) {
+  output.instances.forEach(function (instance) {
     let tags = '';
     Object.keys(instance.tags).forEach(function (key) {
       if (key === 'uid') {
@@ -27,6 +15,6 @@ module.exports = function (context, instances) {
     });
     tabl.push([tags, instance.summary.pass, instance.summary.fail, instance.summary.total, instance.reportFile || '']);
   });
-  tabl.push(['TOTALS', totals.pass, totals.fail, totals.total]);
+  tabl.push(['TOTALS', output.totals.pass, output.totals.fail, output.totals.total]);
   console.log(tabl.toString());
 };
