+++ lib/flow.js
@@ -8,8 +8,6 @@ var reporter = require('../lib/reporter');
 var Glob = require('glob');
 var path = require('path');
 var moment = require('moment');
-var fs = require('fs');
-
 
 let profile = function profile(cb) {
   var base = this.config.get('profiles:base');
@@ -45,9 +43,6 @@ let reportDir = function reportDir(cb) {
   let tsDirName = moment().format("MM-DD-YYYY/HH-mm-ss");
   let fullReportPath = `${reportOutput}/${tsDirName}`;
   this.config.set('output:reports', fullReportPath);
-  this.emitter.on('master:end', (context, summary) => {
-    fs.writeFileSync(`${this.config.get('output:reports')}/summary.json`, summary);
-  });
   log(`reportDir: ${fullReportPath}`);
   this.instances.forEach(function (instance) {
     instance.conf.reports = fullReportPath;
@@ -165,10 +160,10 @@ let reportFiles = function reportFiles(cb) {
     // set up reporter options
     if (reporter.hasOwnProperty(reporterFromConfig)) {
       log(`reportFiles: we have a handler for ${reporterFromConfig}`);
-      reporter[reporterFromConfig](instance);
+      reporter[reporterFromConfig](this, instance);
     }
     instances.push(instance);
-  });
+  }.bind(this));
   this.instances = instances;
   cb(null, this);
 };
