+++ lib/reporter.js
@@ -1,8 +1,17 @@
 var debug = require('debug');
 var log = debug('nemo:reporter:log');
 var filenamify = require('filenamify');
+var fs = require('fs');
+var util = require('./util');
 
-module.exports.mochawesome = function mochawesome(instance) {
+function masterEnd(context) {
+  context.emitter.once('master:end', (_context, _instances) => {
+    let output = util.formatSummary(_instances);
+    fs.writeFileSync(`${_context.config.get('output:reports')}/summary.json`, JSON.stringify(output, null, '\t'));
+  });
+}
+module.exports.mochawesome = function mochawesome(context, instance) {
+  masterEnd(context);
   if (!instance.conf.reports) {
     log('mochawesome: output.reports was not defined. Fallback to manual mochawesome config');
     return;
@@ -25,7 +34,8 @@ module.exports.mochawesome = function mochawesome(instance) {
   instance.conf.mocha.reporterOptions = Object.assign({}, instance.conf.mocha.reporterOptions, reporterOptions);
 };
 
-module.exports.xunit = function (instance) {
+module.exports.xunit = function (context, instance) {
+  masterEnd(context);
   if (!instance.conf.reports) {
     log('xunit: output.reports was not defined. Fallback to manual xunit config');
     return;
